+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 [unreleased]
+
+#### 1.3.5 / 2022-03-31
 * add more Site Health info for runtime environment
 * update `move_dir()`
 * add `is_virtualbox()`